Feature/admin

+++ sideProject/src/main/java/tw/platform/sideProject/service/TagService.java
@@ -8,7 +8,6 @@ import tw.platform.sideProject.model.Tag;
 import tw.platform.sideProject.repository.TagRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class TagService {
@@ -17,6 +16,7 @@ public class TagService {
     private TagRepository tagRepository;
 
     public String addTags(List<Tag> tags) {
+    	
     	// 新增技能 (Tag)
         Tag javaTag = new Tag();
         javaTag.setTagidm(5);
@@ -33,34 +33,23 @@ public class TagService {
         return "標籤新增成功";
     }
 
-    // 新增標籤
-    @Transactional
-    public void saveTag(Tag tag) {
-        tagRepository.save(tag);
-    }
-
-    // 查詢所有標籤
-    public List<Tag> getAllTags() {
-        return tagRepository.findAll();
-    }
-
-    // 根據ID查詢標籤
-    public Optional<Tag> getTagById(int tagId) {
-        return tagRepository.findById(tagId);
-    }
-
     // 更新標籤資料
+    @Transactional
     public Tag updateTag(int tagId, Tag tagDetails) {
-    	
         return tagRepository.findById(tagId).map(tag -> {
             tag.setTagname(tagDetails.getTagname());
             tag.setTagido(tagDetails.getTagido());
             return tagRepository.save(tag);
-        }).orElseThrow(() -> new RuntimeException("標籤ID不存在"));
+        }).orElseThrow(() -> new RuntimeException("標籤ID不存在 (ID: " + tagId + ")"));
     }
 
     // 刪除標籤
+    @Transactional
     public void deleteTag(int tagId) {
-        tagRepository.deleteById(tagId);
+        if (tagRepository.existsById(tagId)) {
+            tagRepository.deleteById(tagId);
+        } else {
+            throw new RuntimeException("標籤ID不存在 (ID: " + tagId + ")");
+        }
     }
 }
